classified importer

+++ web/modules/custom/etype/modules/etype_classified_importer/src/Controller/ImportClassifiedController.php
@@ -146,10 +146,11 @@ class ImportClassifiedController {
     $query = \Drupal::entityQuery('node');
     $query->condition('type', 'classified_ad');
     $tids = $query->execute();
-    $storage_handler = \Drupal::entityTypeManager()->getStorage('node');
-    $entities = $storage_handler->loadMultiple($tids);
-    var_dump($entities);exit;
-    $storage_handler->delete($entities);
+    if (count($tids) > 0) {
+      $storage_handler = \Drupal::entityTypeManager()->getStorage('node');
+      $entities = $storage_handler->loadMultiple($tids);
+      $storage_handler->delete($entities);
+    }
 
     return ['#markup' => ''];
